refactor(solver): change Diagnostics candidates_evaluated from u32 to u64
The Diagnostics struct's candidates_evaluated field was changed from u32 to u64 to handle larger counts of candidates evaluated by the solver. Added a Default derive to Diagnostics and updated tests accordingly to use Diagnostics::default() for cleaner initialization. Updated documentation to reflect the type change.

+++ wildside-core/src/solver.rs
@@ -80,12 +80,12 @@ impl SolveRequest {
 /// assert_eq!(diagnostics.candidates_evaluated, 150);
 /// ```
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
 pub struct Diagnostics {
     /// Time taken to produce the solution.
     pub solve_time: std::time::Duration,
     /// Number of candidate POIs evaluated by the solver.
-    pub candidates_evaluated: u32,
+    pub candidates_evaluated: u64,
 }
 
 /// Response from a successful solve.
